Add maps, scripts, and walkthroughs
Maps are titled "dungeon-2017_V<version>.png" and should be used with the version you are working with. They have been copied from the original map (a real work of art) so they are easier to read. Scripts will be found in "Scripted Moves.txt" and contain spoilers! Be warned! The walk-through files also contain spoilers!